+++ src/cryptography/x509/general_name.py
@@ -46,7 +46,15 @@ class GeneralName(metaclass=abc.ABCMeta):
 
 
 class RFC822Name(GeneralName):
+  <<<<<<< dependabot/github_actions/actions/upload-artifact-v2.2.2
     def __init__(self, value: str) -> None:
+  =======
+  <<<<<<< circleci-project-setup
+    def __init__(self, value: str) -> None:
+  =======
+    def __init__(self, value: str):
+  >>>>>>> 3.4.x
+  >>>>>>> 3.4.x
         if isinstance(value, str):
             try:
                 value.encode("ascii")
@@ -94,7 +102,15 @@ class RFC822Name(GeneralName):
 
 
 class DNSName(GeneralName):
+  <<<<<<< dependabot/github_actions/actions/upload-artifact-v2.2.2
     def __init__(self, value: str) -> None:
+  =======
+  <<<<<<< circleci-project-setup
+    def __init__(self, value: str) -> None:
+  =======
+    def __init__(self, value: str):
+  >>>>>>> 3.4.x
+  >>>>>>> 3.4.x
         if isinstance(value, str):
             try:
                 value.encode("ascii")
@@ -136,7 +152,15 @@ class DNSName(GeneralName):
 
 
 class UniformResourceIdentifier(GeneralName):
+  <<<<<<< dependabot/github_actions/actions/upload-artifact-v2.2.2
     def __init__(self, value: str) -> None:
+  =======
+  <<<<<<< circleci-project-setup
+    def __init__(self, value: str) -> None:
+  =======
+    def __init__(self, value: str):
+  >>>>>>> 3.4.x
+  >>>>>>> 3.4.x
         if isinstance(value, str):
             try:
                 value.encode("ascii")
@@ -180,7 +204,15 @@ class UniformResourceIdentifier(GeneralName):
 
 
 class DirectoryName(GeneralName):
+  <<<<<<< dependabot/github_actions/actions/upload-artifact-v2.2.2
     def __init__(self, value: Name) -> None:
+  =======
+  <<<<<<< circleci-project-setup
+    def __init__(self, value: Name) -> None:
+  =======
+    def __init__(self, value: Name):
+  >>>>>>> 3.4.x
+  >>>>>>> 3.4.x
         if not isinstance(value, Name):
             raise TypeError("value must be a Name")
 
@@ -207,7 +239,15 @@ class DirectoryName(GeneralName):
 
 
 class RegisteredID(GeneralName):
+  <<<<<<< dependabot/github_actions/actions/upload-artifact-v2.2.2
     def __init__(self, value: ObjectIdentifier) -> None:
+  =======
+  <<<<<<< circleci-project-setup
+    def __init__(self, value: ObjectIdentifier) -> None:
+  =======
+    def __init__(self, value: ObjectIdentifier):
+  >>>>>>> 3.4.x
+  >>>>>>> 3.4.x
         if not isinstance(value, ObjectIdentifier):
             raise TypeError("value must be an ObjectIdentifier")
 
@@ -234,7 +274,23 @@ class RegisteredID(GeneralName):
 
 
 class IPAddress(GeneralName):
+  <<<<<<< dependabot/github_actions/actions/upload-artifact-v2.2.2
     def __init__(self, value: _IPADDRESS_TYPES) -> None:
+  =======
+  <<<<<<< circleci-project-setup
+    def __init__(self, value: _IPADDRESS_TYPES) -> None:
+  =======
+    def __init__(
+        self,
+        value: typing.Union[
+            ipaddress.IPv4Address,
+            ipaddress.IPv6Address,
+            ipaddress.IPv4Network,
+            ipaddress.IPv6Network,
+        ],
+    ):
+  >>>>>>> 3.4.x
+  >>>>>>> 3.4.x
         if not isinstance(
             value,
             (
@@ -273,7 +329,15 @@ class IPAddress(GeneralName):
 
 
 class OtherName(GeneralName):
+  <<<<<<< dependabot/github_actions/actions/upload-artifact-v2.2.2
     def __init__(self, type_id: ObjectIdentifier, value: bytes) -> None:
+  =======
+  <<<<<<< circleci-project-setup
+    def __init__(self, type_id: ObjectIdentifier, value: bytes) -> None:
+  =======
+    def __init__(self, type_id: ObjectIdentifier, value: bytes):
+  >>>>>>> 3.4.x
+  >>>>>>> 3.4.x
         if not isinstance(type_id, ObjectIdentifier):
             raise TypeError("type_id must be an ObjectIdentifier")
         if not isinstance(value, bytes):