+++ CHANGELOG.rst
@@ -1,8 +1,12 @@
 Changelog
 =========
 
+  <<<<<<< dependabot/github_actions/actions/upload-artifact-v2.2.2
   <<<<<<< 3.1.x
   =======
+  =======
+  <<<<<<< circleci-project-setup
+  >>>>>>> 3.4.x
 .. _v35-0-0:
 
 35.0.0 - `main`_
@@ -15,6 +19,7 @@ Changelog
   :class:`~cryptography.hazmat.primitives.ciphers.algorithms.SM4`,
   when using OpenSSL 1.1.1. These algorithms are provided for compatibility
   in regions where they may be required, and are not generally recommended.
+  <<<<<<< dependabot/github_actions/actions/upload-artifact-v2.2.2
 * We now ship ``manylinux_2_24`` wheels, in addition to our ``manylinux2010``
   and ``manylinux2014`` wheels.
 
@@ -26,6 +31,11 @@ Changelog
 * Updated Windows, macOS, and ``manylinux`` wheels to be compiled with
   OpenSSL 1.1.1k.
 
+  =======
+
+  =======
+  >>>>>>> 3.4.x
+  >>>>>>> 3.4.x
 .. _v3-4-6:
 
 3.4.6 - 2021-02-16